Decompose toolpath path, cross hair and page template canvas overlays
